Modernised/Consistentised the UI
I just put a banner and put some new css code to make the pages better. Also now the pages are well connected then before. Also the oldest segment of the site in blog directory  got remade with consistent UI as the Main Page. Also now a 404 Page has been added to show the error in case the page is deleted or directory not exists.